refactor(frontend): InvitationCodesTable: InvitationCodesPage delegates logic to InvitationCodesTable component
using new component "InvitationCodesTable" for table logic on InvitationCodesPage
