+++ resources/js/Pages/InvitationCodes/InvitationCodesPage.tsx
@@ -1,38 +1,24 @@
-import CheckMark from '@/Components/Icon/CheckMark';
-import CrossMark from '@/Components/Icon/CrossMark';
-import { InvitationCodeAddButton } from '@/Components/InvitationCodes/InvitationCodeAddButton';
-import { InvitationCodeDeleteButton } from '@/Components/InvitationCodes/InvitationCodeDeleteButton';
-import InvitationCodeInvitationLinkButton from '@/Components/InvitationCodes/InvitationCodeInvitationLinkButton';
-import { UserDeleteButton } from '@/Components/Users/UserDeleteButton';
-
-import Table, { Row } from '@/Components/Util/Table';
+import { InvitationCodesTable } from '@/Components/InvitationCodes/InvitationCodesTable';
 import AuthenticatedLayout from '@/Layouts/AuthenticatedLayout';
 import { InvitationCode } from '@/types/invitation-code.model';
 import { PageProps } from '@/types/page-props.type';
-import { UserMin } from '@/types/user-min.model';
 import { User } from '@/types/user.model';
 import { Head } from '@inertiajs/react';
-import { ReactNode } from 'react';
 
 export default function InvitationCodesPage({
   auth,
   invitationCodes,
 }: PageProps<{ invitationCodes: InvitationCode[] }>) {
+  const admin = auth.user;
   const showAddCodeButton =
     !hasActiveAdminAlreadyCreatedOneStillUnusedInvitationCode(
       invitationCodes,
-      auth.user
+      admin
     );
-
-  const addCodeButtonRow: Row | undefined = showAddCodeButton
-    ? createAddCodeButtonRow()
-    : undefined;
-
-  const tableRows = createRows(invitationCodes, addCodeButtonRow);
 
   return (
     <AuthenticatedLayout
-      user={auth.user}
+      user={admin}
       header={
         <h2 className="font-semibold text-xl text-gray-800 leading-tight">
           User & Einladungen
@@ -43,16 +29,10 @@ export default function InvitationCodesPage({
 
       <div className="py-12">
         <div className="max-w-7xl mx-auto sm:px-6 lg:px-8 space-y-6">
-          <Table
-            headerTitles={[
-              'code',
-              'Anwendbar',
-              'Eingeladen',
-              'Angelegt Durch',
-              'Löschen',
-            ]}
-            rowItems={tableRows}
-          ></Table>
+          <InvitationCodesTable
+            showAddCodeButton={showAddCodeButton}
+            invitationCodes={invitationCodes}
+          />
         </div>
       </div>
     </AuthenticatedLayout>
@@ -66,46 +46,3 @@ const hasActiveAdminAlreadyCreatedOneStillUnusedInvitationCode = (
   invitationCodes
     .filter(invitationCode => invitationCode.admin_id === admin.id)
     .filter(invitationCode => invitationCode.editor_id === null).length > 0;
-
-const createRows = (
-  invitationCodes: InvitationCode[],
-  addCodeButtonRow: Row | undefined
-): Row[] => [
-  ...(addCodeButtonRow ? [addCodeButtonRow] : []),
-  ...invitationCodes.map(({ id, code, editor, admin }) => ({
-    key: code,
-    nodes: [
-      code,
-      editor ? <CrossMark /> : <CheckMark />,
-      editor ? (
-        userMail(editor)
-      ) : (
-        <InvitationCodeInvitationLinkButton code={code} />
-      ),
-      userMail(admin),
-      createDeleteButton(id, editor),
-    ],
-  })),
-];
-
-const createAddCodeButtonRow = () => ({
-  key: 'InvitationCodeAddButton',
-  nodes: [<InvitationCodeAddButton key={'InvitationCodeAddButton'} />],
-});
-
-const createDeleteButton = (codeId: number, editor: UserMin | null) =>
-  editor ? (
-    <UserDeleteButton key={`delete_user_${editor.id}`} user={editor} />
-  ) : (
-    <InvitationCodeDeleteButton key={`delete_code_${codeId}`} id={codeId} />
-  );
-
-const userMail = (user: UserMin): ReactNode => (
-  <a
-    className="text-blue-800"
-    href={`mailto:${user.email}`}
-    aria-label={`Email ${user.name}`}
-  >
-    {user.name}
-  </a>
-);
